refactor(turso-drizzle-adapter): 公開APIと内部翻訳層を分離して構造化
公開ファイルと内部実装がフラットに並んでいて、どれが外部に公開されているかパッと見分からなかった。`drizzle-orm/sqlite-proxy` への翻訳層 (proxy-execute / proxy-transaction) は本パッケージ内部からしか呼ばれないため `sqlite-proxy/` サブディレクトリに分離し、`src/` 直下を公開 API のみに揃える。

併せて `index.ts` を純粋な re-export に整理し、`SqliteRunResult` 型を `sqlite-run-result.ts` に切り出して「1ファイル1export」のコーディング規約に揃えた。

CLAUDE.md を新規作成し、公開 API / 内部構造 / 暫定アダプタである旨 / Phase 4 拡張余地を文書化。
