+++ packages/turso-drizzle-adapter/src/sqlite-proxy/proxy-transaction.test.ts
@@ -1,6 +1,6 @@
 import type { Database } from "@tursodatabase/database";
 import { afterEach, beforeEach, describe, expect, test } from "vitest";
-import { createTursoDatabaseHandle } from "./create-turso-database-handle";
+import { createTursoDatabaseHandle } from "../create-turso-database-handle";
 import { proxyTransaction } from "./proxy-transaction";
 
 describe("proxyTransaction", () => {
